DailyShare/requirements.txt
selenium
webdriver-manager
beautifulsoup4
pyyaml
gitpython
requests
transformers
torch
